Adjusting keybindings
Rational: Using ',' as a "leader" is not working out for me.
Since changing from sneak ro leap I decided to stay with the
native vim "f t F T ; ," which is s problem for ','.  I also
have to think too much to use it.  So, replacing ',' with '\'
mostly because I under utilize '\' and it is completely unbound
in vi.  In nvim/vim '<C-\><C-n>' is used to get into normal-mode
from terminal-mode.

- <space>n now cycles thru absolute numbering before relative ordering
- replace ',' with '\' for blackhole and clipboard keymaps
- added '\D' & '\C' normal mode keymaps for "_ register
- added '\sD' & '\sC' normal mode keymaps for "+ register
- DAP prefix now <bslash><bslash>
- replaced '\\' with '<bslash>' in keybindings
- keep '\\' for whichkey labels